Cleanup the processors with a shared class
Probably think about moving this to sst-plugininfra one day
actually but for now park it here.
